+++ data/scripts/talkactions/player/leave_house.lua
@@ -1,6 +1,11 @@
 local leaveHouse = TalkAction("!leavehouse")
 
 function leaveHouse.onSay(player, words, param)
+	if configManager.getBoolean(configKeys.CYCLOPEDIA_HOUSE_AUCTION) then
+		player:sendTextMessage(MESSAGE_FAILURE, "Command have been disabled by the administrator.")
+		return true
+	end
+
 	local playerPosition = player:getPosition()
 	local playerTile = Tile(playerPosition)
 	local house = playerTile and playerTile:getHouse()